refactor: move server actions file to _actions
